refactor(common/spotify/mocks): move `formatted` and `spotify` mocks to one file
