+++ libs/common/src/spotify/mocks/index.ts
@@ -1,9 +1,5 @@
-export * from './formatted-artist.mock'
-export * from './formatted-track.mock'
-export * from './formatted-profile.mock'
-export * from './formatted-device.mock'
-export * from './spotify-track.mock'
-export * from './spotify-artist.mock'
-export * from './spotify-profile.mock'
-export * from './spotify-device.mock'
+export * from './artist.mock'
+export * from './track.mock'
+export * from './profile.mock'
+export * from './device.mock'
 export * from './playback-state.mock'
