+++ libs/common/src/spotify/mocks/profile.mock.ts
@@ -1,4 +1,4 @@
-import { SpotifyProfile } from '../types'
+import { FormattedProfile, SpotifyProfile } from '../types'
 
 export const spotifyProfileMock: SpotifyProfile = {
   country: 'US',
@@ -39,4 +39,30 @@ export const spotifyProfileMock: SpotifyProfile = {
   product: 'premium',
   type: 'user',
   uri: 'spotify:user:spotify-user',
+}
+
+export const formattedProfileMock: FormattedProfile = {
+  country: 'US',
+  displayName: 'Spotify User',
+  email: 'spotify-user@example.com',
+  followers: 0,
+  id: 'spotify-user',
+  uri: 'spotify:user:spotify-user',
+  images: [
+    {
+      height: 640,
+      url: 'https://i.scdn.co/image/ab6761610000e5eb0f08b008da9d91574819c92e',
+      width: 640,
+    },
+    {
+      height: 320,
+      url: 'https://i.scdn.co/image/ab676161000051740f08b008da9d91574819c92e',
+      width: 320,
+    },
+    {
+      height: 160,
+      url: 'https://i.scdn.co/image/ab6761610000f1780f08b008da9d91574819c92e',
+      width: 160,
+    },
+  ],
 }
