--- libs/common/src/spotify/mocks/spotify-device.mock.ts
@@ -1,16 +1,0 @@
-import { SpotifyDevice } from '../types'
-
-export const spotifyDeviceMock: SpotifyDevice = {
-  id: 'id',
-  is_active: true,
-  is_private_session: false,
-  is_restricted: false,
-  name: 'name',
-  type: 'type',
-  volume_percent: 100,
-}
-
-export const spotifyDevicesMock: SpotifyDevice[] = Array.from(
-  { length: 5 },
-  () => spotifyDeviceMock
-)
